+++ src/main/java/kitchenpos/application/dto/response/OrderLineItemResponse.java
@@ -2,7 +2,7 @@ package kitchenpos.application.dto.response;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderLineItem;
 
 public class OrderLineItemResponse {
