+++ src/main/java/kitchenpos/application/dto/response/OrderResponse.java
@@ -3,7 +3,7 @@ package kitchenpos.application.dto.response;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Order;
+import kitchenpos.order.domain.Order;
 
 public class OrderResponse {
