+++ src/main/java/kitchenpos/order/domain/OrderLineItem.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.order.domain;
 
 import static java.util.Objects.nonNull;
